+++ camera_v0.0.3.py
@@ -1,37 +1,38 @@
-import turtle
-import numpy as np
 import cv2
+import time
 
 cap = cv2.VideoCapture(0)
 
-fourcc = cv2.VideoWriter_fourcc(*'XVID')
+font = cv2.FONT_HERSHEY_SIMPLEX
+currentFrameNumber = 1
 
-#fereastra mica calculator
-#out = cv2.VideoWriter('output.avi',fourcc, 20.0, (640,480))
-
-#smmartphone
-#out = cv2.VideoWriter('output.avi',fourcc, 20.0, (480,800))
-
-#test
-#out = cv2.VideoWriter('output.avi',fourcc, 20.0, (1080,1080))
-
-#test_2
-out = cv2.VideoWriter('output.avi',fourcc, 65535, (640,480))
+t0 = time.time()
 
 while(cap.isOpened()):
     ret, frame = cap.read()
-    if ret==True:
-        frame = cv2.flip(frame,1)
+    
+    if ret == True:
+        frame = cv2.flip(frame, 1)
 
+        windowName = 'video-window'
+
+        cv2.putText(frame, "%s frames" % str(currentFrameNumber), (10,30), font, 1, (200,0,0), 2)
         
-        out.write(frame)
+        currentTimestamp = time.time()
+        delta = currentTimestamp - t0
+        cv2.putText(frame, "%s seconds" % ('%.2f' % delta), (10,60), font, 1, (200,210,0), 2)
 
-        cv2.imshow('frame',frame)
-        if cv2.waitKey(1) & 0xFF == ord('q'):
+        fps = currentFrameNumber / delta
+        cv2.putText(frame, "%s fps" % ('%.2f' % fps), (10,90), font, 1, (0,200,0), 2)
+
+        cv2.imshow(windowName, frame)     
+        
+        if cv2.waitKey(1) & 0xFF == ord('x'):
             break
+        
+        currentFrameNumber = currentFrameNumber + 1
     else:
         break
 
 cap.release()
-out.release()
 cv2.destroyAllWindows()